requirements.txt
celery
certifi
cryptography
Django
django-celery-beat
django-constance
django-cors-headers
djangorestframework
dotenv
djangorestframework-simplejwt
gunicorn
ipython
redis
requests
django_redis
mysqlclient
mysql-connector-python
chardet
numpy